+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - API changes:
   - Introduced `forEvery` to run a Shake action over a pattern of files when they change.
   - Removed `buildHtmlMulti` (use `forEvery` with `buildHtml` instead)
+  - Exposed `Rib.Shake.writeFileCached`
 
 ## 0.6.0.0