+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - mmark: 0.0.7.2
   - megaparsec: 0.8
   - clay: 0.14
+  - shake: 0.8.15
 - MMark, extensions removed:
   - `ghcSyntaxHighlighter`: we already have `skylighting` (which supports more parsers than Haskell)
   - `obfuscateEmail`: requires JS, which is not documented.